pyproject.toml
[tool.poetry]
name = "linkedin-job-matcher"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
pandas = "^2.2.2"
duckdb = "^1.0.0"
pyarrow = "^16.1.0"
pyyaml = "^6.0.1"
sentence-transformers = "^3.0.1"
faiss-cpu = "^1.8.0.post1"
pymupdf = "^1.24.7"
pdfplumber = "^0.11.1"


[tool.poetry.group.development.dependencies]
pytest = "^8.2.2"
python-dotenv = "^1.0.1"
nox = "^2024.4.15"
black = "^24.4.2"
autoflake = "^2.3.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
include = '\.py$'
